Reorganise file structure for utils func in pipeline
- Migrated columns.py and extract.py to utils directory
