+++ content-optimization/src/content_optimization/utils/columns.py
